Changes to Inventory Manager, more Tile Manager Gardening Work
Inventory Manager now has its own encapsulated function for handling its data to and from the Player Persistence Scriptable Object.
Gardening routine now checks for a valid diggable tile before digging.

+++ Assets/Scripts/Input_Manager.cs
@@ -256,8 +256,8 @@ public class Input_Manager : MonoBehaviour
             scrollTimer++;
             if(scrollTimer == 4)
             {
-                if(scrollingDown) invManager.toggleSelection(-1);
-                else if (scrollingUp) invManager.toggleSelection(1);
+                if(scrollingDown) invManager.toggleSelection((sbyte)-1);
+                else if (scrollingUp) invManager.toggleSelection((sbyte)1);
                 scrollTimer = 0;
             }
         }
